Extract visitor classes to separate files
