+++ Tree.cpp
@@ -1,5 +1,7 @@
 #include "Tree.h"
 #include "visitor.h"
+#include "evaluator.h"
+#include "optimizer.h"
 
 void Tree::optimize() {
     auto optimizer = new OptimizationVisitor();
